+++ src/main/kotlin/ch/icken/processor/PanacheCompanionBaseProcessor.kt
@@ -53,94 +53,83 @@ class PanacheCompanionBaseProcessor(
 
         val className = ksClass.toClassName()
         val columnsObjectClassName = ClassName(packageName, classSimpleName + SUFFIX_OBJECT_COLUMNS)
-
+        val companionClassName = className.nestedClass(CLASS_NAME_COMPANION)
         val idClassName = ksClass.getAllProperties()
             .filter { it.hasAnnotation(JakartaPersistenceId) }
             .single()
             .type.resolve()
             .toClassName()
 
-        val expressionParameterType = LambdaTypeName.get(
+        val expressionType = ExpressionClassName.plusParameter(columnsObjectClassName)
+        val expressionParameterLambdaType = LambdaTypeName.get(
             receiver = columnsObjectClassName,
-            returnType = BooleanExpressionClassName
+            returnType = expressionType
         )
         val queryComponentType = QueryComponentClassName
             .plusParameter(className)
             .plusParameter(idClassName)
-        val groupComponentParameterType = LambdaTypeName.get(
-            receiver = queryComponentType,
-            returnType = queryComponentType
-        )
+            .plusParameter(columnsObjectClassName)
 
-        val whereReceiver = className.nestedClass("Companion")
         //region where
         val where = FunSpec.builder(FUNCTION_NAME_WHERE)
             .addModifiers(KModifier.INLINE)
-            .receiver(whereReceiver)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
+            .receiver(companionClassName)
+            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterLambdaType)
             .returns(queryComponentType)
             .addStatement("return %M($PARAM_NAME_EXPRESSION(%T))",
                 MemberName(QueryComponentClassName.packageName, FUNCTION_NAME_WHERE), columnsObjectClassName)
             .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_WHERE$classSimpleName"))
-        val whereGroup = FunSpec.builder(FUNCTION_NAME_WHERE_GROUP)
-            .receiver(whereReceiver)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
-            .addParameter(PARAM_NAME_GROUP_COMPONENT, groupComponentParameterType)
-            .returns(queryComponentType)
-            .addStatement("return %M($PARAM_NAME_EXPRESSION(%T), $PARAM_NAME_GROUP_COMPONENT)",
-                MemberName(QueryComponentClassName.packageName, FUNCTION_NAME_WHERE_GROUP), columnsObjectClassName)
-            .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_WHERE_GROUP$classSimpleName"))
         //endregion
 
-        //region and
+        //region and, or
         val and = FunSpec.builder(FUNCTION_NAME_AND)
             .addModifiers(KModifier.INLINE)
             .receiver(queryComponentType)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
-            .returns(AndQueryComponentClassName.plusParameter(className).plusParameter(idClassName))
+            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterLambdaType)
+            .returns(AndQueryComponentClassName.plusParameter(className)
+                .plusParameter(idClassName).plusParameter(columnsObjectClassName))
             .addStatement("return $FUNCTION_NAME_AND($PARAM_NAME_EXPRESSION(%T))", columnsObjectClassName)
             .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_AND$classSimpleName"))
-        val andGroup = FunSpec.builder(FUNCTION_NAME_AND_GROUP)
-            .receiver(queryComponentType)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
-            .addParameter(PARAM_NAME_GROUP_COMPONENT, groupComponentParameterType)
-            .returns(queryComponentType)
-            .addStatement("return $FUNCTION_NAME_AND_GROUP($PARAM_NAME_EXPRESSION(%T), $PARAM_NAME_GROUP_COMPONENT)",
-                columnsObjectClassName)
-            .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_AND_GROUP$classSimpleName"))
-        //endregion
-
-        //region or
         val or = FunSpec.builder(FUNCTION_NAME_OR)
             .addModifiers(KModifier.INLINE)
             .receiver(queryComponentType)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
-            .returns(OrQueryComponentClassName.plusParameter(className).plusParameter(idClassName))
+            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterLambdaType)
+            .returns(OrQueryComponentClassName.plusParameter(className)
+                .plusParameter(idClassName).plusParameter(columnsObjectClassName))
             .addStatement("return $FUNCTION_NAME_OR($PARAM_NAME_EXPRESSION(%T))", columnsObjectClassName)
             .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_OR$classSimpleName"))
-        val orGroup = FunSpec.builder(FUNCTION_NAME_OR_GROUP)
-            .receiver(queryComponentType)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
-            .addParameter(PARAM_NAME_GROUP_COMPONENT, groupComponentParameterType)
-            .returns(queryComponentType)
-            .addStatement("return $FUNCTION_NAME_OR_GROUP($PARAM_NAME_EXPRESSION(%T), $PARAM_NAME_GROUP_COMPONENT)",
-                columnsObjectClassName)
-            .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_OR_GROUP$classSimpleName"))
+        //endregion
+
+        //region andExpression, orExpression
+        val andExpression = FunSpec.builder(FUNCTION_NAME_AND)
+            .addModifiers(KModifier.INLINE)
+            .receiver(expressionType)
+            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterLambdaType)
+            .returns(expressionType)
+            .addStatement("return $FUNCTION_NAME_AND($PARAM_NAME_EXPRESSION(%T))", columnsObjectClassName)
+            .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_AND_EXPRESSION$classSimpleName"))
+        val orExpression = FunSpec.builder(FUNCTION_NAME_OR)
+            .addModifiers(KModifier.INLINE)
+            .receiver(expressionType)
+            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterLambdaType)
+            .returns(expressionType)
+            .addStatement("return $FUNCTION_NAME_OR($PARAM_NAME_EXPRESSION(%T))", columnsObjectClassName)
+            .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_OR_EXPRESSION$classSimpleName"))
         //endregion
 
         //region count, delete, find, stream
         val count = FunSpec.builder(FUNCTION_NAME_COUNT)
             .addModifiers(KModifier.INLINE)
-            .receiver(whereReceiver)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
+            .receiver(companionClassName)
+            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterLambdaType)
             .returns(LongClassName)
             .addStatement("return $FUNCTION_NAME_WHERE($PARAM_NAME_EXPRESSION).$FUNCTION_NAME_COUNT()")
             .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_COUNT$classSimpleName"))
 
         val delete = FunSpec.builder(FUNCTION_NAME_DELETE)
             .addModifiers(KModifier.INLINE)
-            .receiver(whereReceiver)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
+            .receiver(companionClassName)
+            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterLambdaType)
             .returns(LongClassName)
             .addStatement("return $FUNCTION_NAME_WHERE($PARAM_NAME_EXPRESSION).$FUNCTION_NAME_DELETE()")
             .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_DELETE$classSimpleName"))
@@ -148,16 +137,16 @@ class PanacheCompanionBaseProcessor(
         val findReturns = PanacheQueryClassName.plusParameter(className)
         val find = FunSpec.builder(FUNCTION_NAME_FIND)
             .addModifiers(KModifier.INLINE)
-            .receiver(whereReceiver)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
+            .receiver(companionClassName)
+            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterLambdaType)
             .returns(findReturns)
             .addStatement("return $FUNCTION_NAME_WHERE($PARAM_NAME_EXPRESSION).$FUNCTION_NAME_FIND()")
             .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_FIND$classSimpleName"))
         val findSorted = FunSpec.builder(FUNCTION_NAME_FIND)
             .addModifiers(KModifier.INLINE)
-            .receiver(whereReceiver)
+            .receiver(companionClassName)
             .addParameter(PARAM_NAME_SORT, SortClassName)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
+            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterLambdaType)
             .returns(findReturns)
             .addStatement("return $FUNCTION_NAME_WHERE($PARAM_NAME_EXPRESSION).$FUNCTION_NAME_FIND($PARAM_NAME_SORT)")
             .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_FIND_SORTED$classSimpleName"))
@@ -165,16 +154,16 @@ class PanacheCompanionBaseProcessor(
         val streamReturn = StreamClassName.plusParameter(className)
         val stream = FunSpec.builder(FUNCTION_NAME_STREAM)
             .addModifiers(KModifier.INLINE)
-            .receiver(whereReceiver)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
+            .receiver(companionClassName)
+            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterLambdaType)
             .returns(streamReturn)
             .addStatement("return $FUNCTION_NAME_WHERE($PARAM_NAME_EXPRESSION).$FUNCTION_NAME_STREAM()")
             .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_STREAM$classSimpleName"))
         val streamSorted = FunSpec.builder(FUNCTION_NAME_STREAM)
             .addModifiers(KModifier.INLINE)
-            .receiver(whereReceiver)
+            .receiver(companionClassName)
             .addParameter(PARAM_NAME_SORT, SortClassName)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
+            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterLambdaType)
             .returns(streamReturn)
             .addStatement("return $FUNCTION_NAME_WHERE($PARAM_NAME_EXPRESSION).$FUNCTION_NAME_STREAM($PARAM_NAME_SORT)")
             .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_STREAM_SORTED$classSimpleName"))
@@ -183,15 +172,15 @@ class PanacheCompanionBaseProcessor(
         //region single
         val single = FunSpec.builder(FUNCTION_NAME_SINGLE)
             .addModifiers(KModifier.INLINE)
-            .receiver(whereReceiver)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
+            .receiver(companionClassName)
+            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterLambdaType)
             .returns(className)
             .addStatement("return $FUNCTION_NAME_WHERE($PARAM_NAME_EXPRESSION).getSingle()")
             .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_SINGLE$classSimpleName"))
         val singleSafe = FunSpec.builder(FUNCTION_NAME_SINGLE_SAFE)
             .addModifiers(KModifier.INLINE)
-            .receiver(whereReceiver)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
+            .receiver(companionClassName)
+            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterLambdaType)
             .returns(PanacheSingleResultClassName.plusParameter(WildcardTypeName.producerOf(className)))
             .addStatement("return $FUNCTION_NAME_WHERE($PARAM_NAME_EXPRESSION).getSingleSafe()")
             .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_SINGLE_SAFE$classSimpleName"))
@@ -201,22 +190,22 @@ class PanacheCompanionBaseProcessor(
         val multipleReturns = ListClassName.plusParameter(className)
         val multiple = FunSpec.builder(FUNCTION_NAME_MULTIPLE)
             .addModifiers(KModifier.INLINE)
-            .receiver(whereReceiver)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
+            .receiver(companionClassName)
+            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterLambdaType)
             .returns(multipleReturns)
             .addStatement("return $FUNCTION_NAME_WHERE($PARAM_NAME_EXPRESSION).getMultiple()")
             .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_MULTIPLE$classSimpleName"))
         val multipleSorted = FunSpec.builder(FUNCTION_NAME_MULTIPLE)
             .addModifiers(KModifier.INLINE)
-            .receiver(whereReceiver)
+            .receiver(companionClassName)
             .addParameter(PARAM_NAME_SORT, SortClassName)
-            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterType)
+            .addParameter(PARAM_NAME_EXPRESSION, expressionParameterLambdaType)
             .returns(multipleReturns)
             .addStatement("return $FUNCTION_NAME_WHERE($PARAM_NAME_EXPRESSION).getMultiple($PARAM_NAME_SORT)")
             .addAnnotation(jvmNameAnnotation("$FUNCTION_NAME_MULTIPLE_SORTED$classSimpleName"))
         //endregion
 
-        val functions = listOf(where, whereGroup, and, andGroup, or, orGroup,
+        val functions = listOf(where, and, or, andExpression, orExpression,
             count, delete, find, findSorted, stream, streamSorted,
             single, singleSafe, multiple, multipleSorted)