+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
+## 0.0.3
+- Remove `whereGroup`, `andGroup`, and `orGroup` functions in favor of expression chaining
+
 ## 0.0.2
-- Add the `@ColumnType` annotation to override the generated `Column<T>`'s type parameter
+- Add the `@ColumnType` annotation to override the generated `Column`'s type parameter
 
 ## 0.0.1
 - Initial release